23f1002424 Project/requirements.txt
fastapi
uvicorn
python-dateutil
requests
scipy
python-dotenv
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.8
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
Levenshtein==0.26.1
openai==1.61.0
pydantic==2.10.6
pydantic_core==2.27.2
python-Levenshtein==0.26.1
RapidFuzz==3.12.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0